storyboard-agent/pyproject.toml
[project]
name = "storyboard-agent"
version = "0.1.0"
description = "Dedicated LangGraph service for Dreamweaver storyboard orchestration"
requires-python = ">=3.12"
dependencies = [
  "deepagents==0.4.1",
  "fastapi==0.116.1",
  "langgraph>=1.0.0,<2.0.0",
  "langgraph-cli[inmem]>=0.4.0",
  "langchain-core>=1.2.10,<2.0.0",
  "langchain-google-genai>=2.0.0",
  "langgraph-checkpoint-postgres>=2.0.0",
  "psycopg[binary]>=3.2",
  "psycopg-pool>=3.2",
  "uvicorn==0.35.0",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deep", "viMax_port", "viMax_port.tests"]
py-modules = ["agent", "linear_agent", "server", "script_ingest_routes"]
